Проведен рефакторинг главного модуля приложения __init__.py
