+++ webapp/modules/IndexPage.py
@@ -1,0 +1,5 @@
+from flask import render_template
+
+def IndexPage():
+    title = 'SimpleMessenger - главная страница'
+    return render_template('index.html', page_title=title)
